DESD-Ibrahim-Final-DESD/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-filter==23.3
django-redis==5.4.0
channels==4.0.0
channels-redis==4.1.0
drf-spectacular==0.26.5
Pillow==10.1.0
gunicorn==21.2.0
Markdown==3.5.1
psycopg2-binary==2.9.9
whitenoise==6.6.0
hiredis==2.2.3
django-db-connection-pool==1.2.2
drf-nested-routers==0.93.4
daphne==4.0.0
firebase-admin

